backend/plant/serializers.py
from rest_framework import serializers
from .models import Plant


class PlantSerializer(serializers.ModelSerializer):
    class Meta:
        model = Plant
        fields = ["id", "name", "dxf_file", "cep", "number", "created_at", "updated_at"]
