backend/requirements.txt
alphashape==1.3.1
asgiref==3.10.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-log==0.4.0
contourpy==1.3.2
cycler==0.12.1
Django==5.1.1
djangorestframework==3.15.2
ezdxf==1.3.2
fonttools==4.60.1
geojson==3.1.0
idna==3.11
kiwisolver==1.4.9
matplotlib==3.8.1
networkx==3.4.2
numpy==1.26.4
opencv-python==4.7.0.72
packaging==25.0
Pillow==9.5.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
reportlab==3.6.12
requests==2.32.5
rtree==1.4.1
scipy==1.15.3
shapely==2.0.3
six==1.17.0
sqlparse==0.5.3
trimesh==4.9.0
typing_extensions==4.15.0
urllib3==2.5.0
